# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Restaurant API",
        "description": "API documentation for the School Restaurant management system.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/meals": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "List all meals",
                "operationId": "b0cec5c3112235d8259f1e5669c3da16",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Meals"
                ],
                "summary": "Create a new meal",
                "operationId": "583fca64e815044fd063c02060a5721e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "meal_type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Beef Steak"
                                    },
                                    "meal_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Grilled beef with vegetables"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Meal created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/meals/{meal}": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "Get meal details",
                "operationId": "48e51ca3c4c6d543db2e2faab8d2df6c",
                "parameters": [
                    {
                        "name": "meal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Meal not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Meals"
                ],
                "summary": "Delete a meal",
                "operationId": "9477155c84bf304d16716adfe134f03f",
                "parameters": [
                    {
                        "name": "meal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Meal deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Meal not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Meals"
                ],
                "summary": "Update a meal",
                "operationId": "3e4a82586dca9fd4891849298ef2f7b6",
                "parameters": [
                    {
                        "name": "meal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "meal_type_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Meal updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Meal not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menu-meals": {
            "get": {
                "tags": [
                    "Menu Meals"
                ],
                "summary": "List meals scheduled for a specific date",
                "operationId": "351f467a64d63d55d34eaedefe6e598d",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "The date to check (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-03-25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Menu Meals"
                ],
                "summary": "Add a meal to the menu for a specific date",
                "operationId": "1696666c1e133d77b1e8b4f2f5615a79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "meal_id",
                                    "reservation_date"
                                ],
                                "properties": {
                                    "meal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Meal added to menu successfully"
                    },
                    "400": {
                        "description": "Meal already in menu for this date"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/menu-meals/{menu_meal}": {
            "delete": {
                "tags": [
                    "Menu Meals"
                ],
                "summary": "Remove a meal from the menu",
                "operationId": "4676ada80994f99e696546a4b3bbd8b9",
                "parameters": [
                    {
                        "name": "menu_meal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meal removed from menu"
                    },
                    "404": {
                        "description": "Menu meal not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}